Add docker env option for session timeout
Confluence has a default session timeout of one hour and no way to adjust this via the application UI.
The confluence docs describe how to modify the `web.xml` in order to adjust the session timeout.
https://confluence.atlassian.com/confkb/how-to-adjust-the-session-timeout-for-confluence-126910597.html
This change adds a new step on container startup that if the env variable `SESSION_TIMEOUT` is set
modifies the appropriated XML values.